trying to figure out slash commands.
