+++ commands/core/ping.js
@@ -5,6 +5,7 @@ module.exports = {
 		.setName('ping')
 		.setDescription('Replies with Pong!'),
 	async execute(interaction) {
+		const x = 6;
 		await interaction.reply('Pong!');
 	},
 };
