+++ deploy-guild-commands.js
@@ -1,8 +1,25 @@
 const { REST } = require('@discordjs/rest');
 const { Routes } = require('discord-api-types/v9');
 const { token, clientId, guildId } = require('./config.json');
-const { getAllFiles } = require('./utils/Utils');
+const fs = require('fs');
+const path = require('path');
 
+
+function getAllFiles(dirPath, arrayOfFiles) {
+	const files = fs.readdirSync(dirPath);
+	arrayOfFiles = arrayOfFiles || [];
+
+	files.forEach(function(file) {
+		if (fs.statSync(dirPath + '/' + file).isDirectory()) {
+			arrayOfFiles = getAllFiles(dirPath + '/' + file, arrayOfFiles);
+		}
+		else {
+			arrayOfFiles.push(path.join('./', dirPath, file));
+		}
+	});
+
+	return arrayOfFiles;
+}
 const commands = [];
 const commandFiles = getAllFiles('./events', []);
