+++ index.js
@@ -1,6 +1,25 @@
 const { Client, Collection, Intents } = require('discord.js');
 const { token } = require('./config.json');
-const { getAllFiles } = require('./utils/Utils');
+const fs = require('fs');
+const path = require('path');
+
+
+function getAllFiles(dirPath, arrayOfFiles) {
+	const files = fs.readdirSync(dirPath);
+	arrayOfFiles = arrayOfFiles || [];
+
+	files.forEach(function(file) {
+		if (fs.statSync(dirPath + '/' + file).isDirectory()) {
+			arrayOfFiles = getAllFiles(dirPath + '/' + file, arrayOfFiles);
+		}
+		else {
+			arrayOfFiles.push(path.join(__dirname, dirPath, file));
+		}
+	});
+
+	return arrayOfFiles;
+}
+
 
 const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
